establish connection between student, class and subject

+++ pages/students/index.js
@@ -3,22 +3,33 @@ import styles from "../../styles/tables.module.css";
 import { useContext, useState, useEffect, useCallback } from "react";
 import { AppContext } from "../../context/app-context";
 import AccessDenied from "../../components/AccessDenied";
+import formStyles from "../../components/Form.module.css";
 
 const StudentsPage = (props) => {
   const { isAuth, user, token } = useContext(AppContext);
 
-  const [students, setStudents] = useState([]);
-  const [loading, setLoading] = useState(true);
+  const [students, setStudents] = useState();
+  const [loading, setLoading] = useState(false);
   const [message, setMessage] = useState(null);
+  const [divisionId, setDivisionId] = useState("");
+  const [optionsList, setOptionsList] = useState([]);
   // ===================================
   // methods
   // ===================================
-  const fetchStudents = useCallback(async () => {
+  const formSubmitHandler = async (e) => {
+    e.preventDefault();
+    if (!divisionId) {
+      setMessage("Select the class from dropdown.");
+      setStudents(null);
+      return;
+    }
     setLoading(true);
     setMessage(null);
     try {
       const res = await fetch(
-        process.env.NEXT_PUBLIC_BACKEND_BASE_URL + "/get-students",
+        process.env.NEXT_PUBLIC_BACKEND_BASE_URL +
+          "/get-students/" +
+          divisionId,
         {
           headers: { Authorization: "Bearer " + token },
         }
@@ -37,14 +48,19 @@ const StudentsPage = (props) => {
       setMessage(error.message);
     }
     setLoading(false);
-  }, [token]);
+  };
 
   const clearStudentsHandler = async (e) => {
+    if (!divisionId) {
+      return;
+    }
     setLoading(true);
     setMessage(null);
     try {
       const res = await fetch(
-        process.env.NEXT_PUBLIC_BACKEND_BASE_URL + "/admin/clear-students",
+        process.env.NEXT_PUBLIC_BACKEND_BASE_URL +
+          "/admin/clear-students/" +
+          divisionId,
         {
           method: "DELETE",
           headers: {
@@ -61,19 +77,45 @@ const StudentsPage = (props) => {
         setMessage(data.message);
       } else {
         setMessage(data.message);
-        setStudents([]);
+        setStudents(null);
       }
     } catch (error) {
       setMessage(error.message);
     }
     setLoading(false);
   };
+
+  const fetchOptionsList = useCallback(async () => {
+    // setLoading(true);
+    // setMessage(null);
+    let endpoint = process.env.NEXT_PUBLIC_BACKEND_BASE_URL + "/get-classes";
+
+    try {
+      const res = await fetch(endpoint, {
+        headers: { Authorization: "Bearer " + token },
+      });
+      if (res.status === 500) {
+        throw new Error("Something went wrong.");
+      }
+      const data = await res.json();
+      if (res.status >= 300) {
+        setMessage(data.message);
+      } else {
+        // setMessage(data.message);
+        setOptionsList(data.divisions);
+      }
+    } catch (error) {
+      // setMessage(error.message);
+    }
+    // setLoading(false);
+  }, [token]);
+
   // ===================================
   // side effects
   // ===================================
   useEffect(() => {
-    fetchStudents();
-  }, [fetchStudents]);
+    fetchOptionsList();
+  }, [fetchOptionsList]);
 
   // ===================================
   // render
@@ -86,7 +128,7 @@ const StudentsPage = (props) => {
   let content;
   let list;
 
-  if (students.length > 0) {
+  if (students && students.length > 0) {
     let studentsList = students.map((student) => {
       return (
         <tr key={student._id} className={styles.row}>
@@ -115,7 +157,7 @@ const StudentsPage = (props) => {
         <tbody>{studentsList}</tbody>
       </table>
     );
-  } else {
+  } else if (students && students.length === 0) {
     content = <p className="msg">No recods found.</p>;
   }
 
@@ -127,12 +169,43 @@ const StudentsPage = (props) => {
     content = <p className="msg">Loading...</p>;
   }
 
+  const divisionOptions = optionsList.map((division) => {
+    return (
+      <option key={division._id} value={division._id}>
+        {division.divisionName}
+      </option>
+    );
+  });
+
   return (
     <div className="main-content">
       <h2 className="title">All Students</h2>
       {content}
+      <form className={formStyles.form} onSubmit={formSubmitHandler}>
+        <div className={formStyles.formControl}>
+          <label htmlFor="divisionId" className={formStyles.label}>
+            Class
+          </label>
+          <select
+            id="divisionId"
+            name="divisionId"
+            value={divisionId}
+            onChange={(e) => {
+              setDivisionId(e.target.value);
+            }}
+          >
+            <option value="">--Please choose the class--</option>
+            {divisionOptions}
+          </select>
+        </div>
+        <div className="buttons">
+          <button className="btn success" type="sumbit">
+            Search
+          </button>
+        </div>
+      </form>
 
-      {students.length > 0 && user.type === "admin" && (
+      {students && students.length > 0 && user.type === "admin" && (
         <div className="buttons">
           <button className="btn danger" onClick={clearStudentsHandler}>
             Clear All
